dest_walls

+++ src/game_map.h
@@ -4,14 +4,14 @@
 
 class GameMap {
   private:
-    double win_height;
-    double win_width;
+    int win_height;
+    int win_width;
     WINDOW * game_win;
     void instantiate_colors();
     void construct_walls();
     void construct_dest_walls();
   public:
-    GameMap(const double &game_screen_max, const double &start_x, const double &start_y);
+    GameMap(const int &game_screen_max, const double &start_x, const double &start_y);
     WINDOW * get_win();
 };
